Fixes compilation errors - moves a function and its test nearest to the code that uses it
